requirements.txt
robin_stocks
pytest
pytest-dotenv